ad/da driver update

+++ demos/hi3093/apps/openamp/soft_spi_driver.h
@@ -4,7 +4,7 @@
 #include <unistd.h>
 
 void spi0_init(void);
-void spi0_adc_receive(uint8_t *rx_buf, size_t length);
-void spi0_dac_transfer(const uint8_t *tx_buf, size_t length);
+void spi0_adc_receive(uint8_t* rx_buf, int length);
+void spi0_dac_transfer(const uint8_t* tx_buf, int length);
 
 #endif
